fix: file check

+++ .codepipeline/codedeploy/application-start.sh
@@ -11,7 +11,15 @@ docker exec $(cd /var/www/cmfive-boilerplate && docker-compose ps -q webapp) php
 docker exec $(cd /var/www/cmfive-boilerplate && docker-compose ps -q webapp) chown -R www-data:www-data /var/www/html
 docker exec $(cd /var/www/cmfive-boilerplate && docker-compose ps -q webapp) php cmfive.php install migrations
 
-sudo rm cache/classdirectory.cache
-sudo rm cache/config.cache
-sudo rm storage/log/*
-sudo rm storage/session/*
+if [ -f "cache/classdirectory.cache" ]; then
+    sudo rm cache/classdirectory.cache
+fi
+if [ -f "cache/config.cache" ]; then
+    sudo rm cache/config.cache
+fi
+if [ -d "storage/log" ]; then
+    sudo rm storage/log/*
+fi
+if [ -d "storage/session" ]; then
+    sudo rm storage/session/*
+fi
